+++ src/main/java/com/proyecto/terranova/dto/ResenaDTO.java
@@ -15,5 +15,5 @@ public class ResenaDTO {
 
     private String cedulaComprador;
 
-    private int idProducto;
+    private Long idProducto;
 }
